Refactor fwd/inv subsystem
